feat: add note creation logic
they will be created through a hidden
form on the createNote div
